+++ Composer/composer.py
@@ -1,10 +1,11 @@
 from Facility import Facility, ActionInformation, DashboardPanel
-from Data import ExperimentDescriptor, ExperimentType
+from Data import ExperimentDescriptor, ExperimentType, NsInfo, Metal
 from .platform_configuration import PlatformConfiguration, TaskDefinition
 from importlib import import_module
 from Helper import Log
 from typing import List
 from Executor.Tasks.Run import Message
+from Interfaces import Management
 
 
 class Composer:
@@ -12,18 +13,43 @@ class Composer:
 
     @classmethod
     def Compose(cls, descriptor: ExperimentDescriptor) -> PlatformConfiguration:
+        def _messageAction(severity: str, message: str) -> ActionInformation:
+            message = ActionInformation()
+            message.TaskName = "Run.Message"
+            message.Order = -9999
+            message.Config = {'Severity': severity, 'Message': message}
+            return message
+
+        def _messageTask(severity: str, message: str) -> TaskDefinition:
+            task = TaskDefinition()
+            task.Task = Message
+            task.Params = {'Severity': severity, 'Message': message}
+            return task
+
         if cls.facility is None:
             cls.facility = Facility()
 
         configuration = PlatformConfiguration()
         configuration.RunParams['Report'] = {'ExperimentName': descriptor.Identifier}
 
-        instantiation = {'NetworkServices': descriptor.NetworkServices}
-        configuration.PreRunParams.update(instantiation)
-        configuration.PostRunParams.update(instantiation)
-
         actions: List[ActionInformation] = []
         panels: List[DashboardPanel] = []
+
+        if len(descriptor.NetworkServices) != 0:
+            sliceManager = Management.SliceManager()
+            for ns in descriptor.NetworkServices:
+                nsId, location = ns
+                try:
+                    nsInfo = NsInfo(nsId, location)
+                    requirements = sliceManager.GetNsdRequirements(nsId)
+                    if requirements is None:
+                        raise RuntimeError("Could not retrieve NSD information")
+                    nsInfo.Requirements = requirements
+                    configuration.NetworkServices.append(nsInfo)
+                except Exception as e:
+                    actions.append(
+                        _messageAction("ERROR",
+                                       f"Exception while obtaining information about network service {nsId}: {e}"))
 
         if descriptor.Type == ExperimentType.MONROE:
             actions.extend(cls.facility.GetMonroeActions())
@@ -31,7 +57,11 @@ class Composer:
             for ue in descriptor.UEs:
                 actions.extend(cls.facility.GetUEActions(ue))
             for testcase in descriptor.TestCases:
-                actions.extend(cls.facility.GetTestCaseActions(testcase))
+                testcaseActions = cls.facility.GetTestCaseActions(testcase)
+                if len(testcaseActions) != 0:
+                    actions.extend(testcaseActions)
+                else:
+                    actions.append(_messageAction("ERROR", f'TestCase "{testcase}" did not generate any actions'))
                 panels.extend(cls.facility.GetTestCaseDashboards(testcase))
 
         actions.sort(key=lambda action: action.Order)  # Sort by Order
@@ -40,14 +70,12 @@ class Composer:
         for action in actions:
             requirements.update(action.Requirements)
 
-            taskDefinition = TaskDefinition()
-            taskDefinition.Params = action.Config
             task = cls.getTaskClass(action.TaskName)
             if task is None:
-                taskDefinition.Task = Message
-                taskDefinition.Params['Severity'] = "ERROR"
-                taskDefinition.Params['Message'] = f"Could not find task {action.TaskName}"
+                taskDefinition = _messageTask("ERROR", f"Could not find task {action.TaskName}")
             else:
+                taskDefinition = TaskDefinition()
+                taskDefinition.Params = action.Config
                 taskDefinition.Task = task
             configuration.RunTasks.append(taskDefinition)