+++ Executor/Tasks/PostRun/decommission.py
@@ -1,20 +1,21 @@
 from Task import Task
 from Helper import Level
+from Data import NsInfo
+from typing import List
 from Interfaces import Management
 
 
 class Decommission(Task):
-    def __init__(self, logMethod, parent, params):
-        super().__init__("Decommission", parent, params, logMethod, None)
+    def __init__(self, logMethod, parent, networkServices):
+        super().__init__("Decommission", parent, {'NetworkServices': networkServices}, logMethod, None)
 
     def Run(self):
         self.Log(Level.INFO, 'Decommision started')
-        networkServices = self.params['NetworkServices']
+        networkServices: List[NsInfo] = self.params['NetworkServices']
         if len(networkServices) != 0:
-            sliceIds = self.params['SliceIds']
-            self.Log(Level.INFO, f"Experiment has instantiated slices: {sliceIds}")
-            for slice in sliceIds:
-                self.Log(Level.INFO, f"Requesting decommision of slice: {slice}")
+            self.Log(Level.INFO, f"Experiment has {len(networkServices)} network services")
+            for ns in networkServices:
+                self.Log(Level.INFO, f"Requesting decommision of NS {ns.Id} with slice ID: {ns.SliceId}")
                 try:
                     # TODO
                     self.Log(Level.INFO, f'Slice decommisioned')