+++ Executor/executor_base.py
@@ -6,22 +6,22 @@ from datetime import datetime, timezone
 from Helper import Serialize
 from .status import Status
 from tempfile import TemporaryDirectory
-from Interfaces import DispatcherApi
+from Interfaces import PortalApi
 
 
 class ExecutorBase(Child):
-    dispatcher: DispatcherApi = None
+    portal: PortalApi = None
 
     def __init__(self, params: Dict, name: str, tempFolder: TemporaryDirectory = None):
-        if self.dispatcher is None:
+        if self.portal is None:
             config = Config()
-            self.dispatcher = DispatcherApi(config.Dispatcher.Host, config.Dispatcher.Port)
+            self.portal = PortalApi(config.Dispatcher.Host, config.Dispatcher.Port)
 
         now = datetime.now(timezone.utc)
         super().__init__(f"{name}{now.strftime('%y%m%d%H%M%S%f')}", tempFolder)
         self.Tag = name
         self.params = params
-        self.Id = params['Id']
+        self.ExecutionId: int = params['ExecutionId']
         self.Created = now
         self.Started = None
         self.Finished = None
@@ -53,7 +53,7 @@ class ExecutorBase(Child):
     def AddMessage(self, msg: str, percent: int = None):
         if percent is not None: self.PerCent = percent
         self.Messages.append(f'[{self.PerCent}%] {msg}')
-        self.dispatcher.UpdateExecutionData(self.Id, percent=self.PerCent, message=msg)
+        self.portal.UpdateExecutionData(self.ExecutionId, percent=self.PerCent, message=msg)
 
     @property
     def LastMessage(self):
@@ -76,7 +76,7 @@ class ExecutorBase(Child):
 
     def Serialize(self) -> Dict:
         data = {
-            'Id': self.Id,
+            'ExecutionId': self.ExecutionId,
             'Name': self.name,
             'Tag': self.Tag,
             'Created': Serialize.DateToString(self.Created),
@@ -94,7 +94,7 @@ class ExecutorBase(Child):
 
     def Save(self):
         data = self.Serialize()
-        path = Serialize.Path(self.Tag, str(self.Id))
+        path = Serialize.Path(self.Tag, str(self.ExecutionId))
         Serialize.Save(data, path)
 
     @classmethod
@@ -102,7 +102,7 @@ class ExecutorBase(Child):
         path = Serialize.Path(tag, id)
         data = Serialize.Load(path)
         tag = data['Tag']
-        params = {'Id': int(id), 'Deserialized': True}
+        params = {'ExecutionId': int(id), 'Deserialized': True}
         if tag == 'PreRunner':
             from .pre_runner import PreRunner
             res = PreRunner(params)
@@ -113,7 +113,8 @@ class ExecutorBase(Child):
             from .post_runner import PostRunner
             res = PostRunner(params)
 
-        res.Id, res.Name, res.LogFile, res.Tag = Serialize.Unroll(data, 'Id', 'Name', 'Log', 'Tag')
+        res.ExecutionId = data.get('ExecutionId', data.get('Id'))  # For compatibility with older serialization
+        res.Name, res.LogFile, res.Tag = Serialize.Unroll(data, 'Name', 'Log', 'Tag')
         res.hasStarted, res.hasFinished = Serialize.Unroll(data, 'HasStarted', 'HasFinished')
         res.Messages, res.PerCent, res.GeneratedFiles = Serialize.Unroll(data, 'Messages', 'PerCent', "GeneratedFiles")
         res.Created = Serialize.StringToDate(data['Created'])