+++ Executor/post_runner.py
@@ -1,5 +1,5 @@
 from typing import Dict
-from .Tasks.PostRun import RequestResults, SaveResults, UpdateExecutionEntry, Decommission
+from .Tasks.PostRun import Decommission, ReleaseResources
 from .executor_base import ExecutorBase
 from tempfile import TemporaryDirectory
 
@@ -11,14 +11,10 @@ class PostRunner(ExecutorBase):
     def Run(self):
         self.SetStarted()
 
-        Decommission(self.Log, self, self.Configuration.PostRunParams).Start()
-        self.AddMessage('Resources decommisioned', 10)
+        Decommission(self.Log, self, self.Configuration.NetworkServices).Start()
+        self.AddMessage('Network services decommisioned', 50)
 
-        RequestResults(self.Log, self).Start()
-        self.AddMessage('Results received', 30)
-        SaveResults(self.Log, self).Start()
-        self.AddMessage('Results stored', 60)
-        UpdateExecutionEntry(self.Log, self).Start()
-        self.AddMessage('Entry updated', 90)
+        ReleaseResources(self.Log, self.ExecutionId, self.Configuration.Requirements, self).Start()
+        self.AddMessage('Released resources', 90)
 
         self.SetFinished(percent=100)