+++ Experiment/execution_tombstone.py
@@ -14,4 +14,5 @@ class Tombstone:
         self.Executor = Executor.Load('Executor', str(self.Id))
         self.PostRunner = Executor.Load('PostRunner', str(self.Id))
         self.Created = Serialize.StringToDate(data['Created'])
+        self.JsonDescriptor = data.get('JsonDescriptor', {})