+++ Helper/dashboard_generator.py
@@ -21,10 +21,10 @@ class DashboardGenerator(RestClient):
         if not self.enabled: return None
 
         body = json.dumps(self.generateData(execution))
-        Log.D(f"Grafana dashboard data (exeuction {execution.Id}): {body}")
-        response = self.httpPost(f"{self.api_url}/dashboards/db", self.headers, body)
+        Log.D(f"Grafana dashboard data (execution {execution.Id}): {body}")
+        response = self.HttpPost(f"{self.api_url}/dashboards/db", self.headers, body)
         if response.status == 200:
-            url = self.responseToJson(response)["url"]
+            url = self.ResponseToJson(response)["url"]
             Log.I(f"Generated Grafana dashboard for execution {execution.Id}: {url}")
             return url
         else: